% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contributions.R
\name{calculate_contributions}
\alias{calculate_contributions}
\title{Calculate individual contributions}
\usage{
calculate_contributions(contributions = NULL)
}
\arguments{
\item{contributions}{A list object (defaults to NULL)}
}
\description{
This function takes a list and calculates individual contributions as a
proportion of total work completed. The output is a dataframe.
}
\examples{
ex <- list("p1" = c(1:13), "p2" = 1, "p3" = c(1, 3, 5, 9))
calculate_contributions(ex)
}
\keyword{dataframe}
