% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_to_text.R
\name{contribution_to_text}
\alias{contribution_to_text}
\title{Print CRediT taxonomy in prose}
\usage{
contribution_to_text(contributor_df)
}
\arguments{
\item{contributor_df}{A dataframe built from contributor_df_builder}
}
\description{
This function takes a contributor dataframe and converts it to prose.
}
\examples{
# A list containing named vectors
ex_1 <- list(
 JVC = 1:13,
 JGP = 1,
 NR  = c(2, 4))

# Simple example
contributor_df_builder(ex_1) |>
contribution_to_text()
}
\keyword{print}
