% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributoR.R
\name{contributor}
\alias{contributor}
\title{Build contributor plot}
\usage{
contributor(contributions = NULL)
}
\arguments{
\item{contributions}{A list object (defaults to NULL)}
}
\description{
This function takes a list and generates a contributions plot.
}
\examples{
# Create example list
ex_1 <- list(
 JVC = 1:13,
 JGP = 1,
 NR  = c(2, 4))

# Plot contributions
contributor(contributions = ex_1)

# Build list inside function
contributor(
  contributions = list(
  "Author 1" = seq(1, 14, by = 3),
  "Author 2" = c(1, 3, 5, 7, 9, 13),
  "Lazy person" = NA,
  "Advisor" = 12)
  )
}
\keyword{plot}
