% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_df_builder.R
\name{contributor_df_builder}
\alias{contributor_df_builder}
\title{Build contributor dataframe}
\usage{
contributor_df_builder(contributions = NULL)
}
\arguments{
\item{contributions}{A list object (defaults to NULL)}
}
\description{
This function takes a list and generates a dataframe that can be sent to
the contributor plotting function
}
\examples{
ex_vec <- c("JVC", "JGP", "NR")
ex_0 <- list(
 "JVC",
 "JGP",
 "NR")
ex_1 <- list(
 JVC = 1:13,
 JGP = 1,
 NR  = c(2, 4))

# Simple example
contributor_df_builder(ex_1)
}
\keyword{dataframe}
